comhpc: Do not update fstab for '*.wic' images
This patch fixes the problem with failing to mount /dev/sda1 under /boot path
when a '*.wic' image is flashed on a device other than /dev/sda e.g.:
/dev/nvme0n1.

The root cause of this problem was an extra 'boot' entry appended by the wic
plugin:  '/dev/sda1     /boot   vfat    defaults        0       0'.

The fix is to prevent the wic plugin from updating the fstab by calling it
with the parameter '--no-fstab-update'.

Signed-off-by: Kamil Dziezyk <kamil.dziezyk@arm.com>